refactor: separate Chatter logic
