+++ src/main/java/com/festp/handlers/ChatHandler.java
@@ -6,6 +6,7 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.player.AsyncPlayerChatEvent;
 
 import com.festp.messaging.Chatter;
+import com.festp.messaging.SpigotMessageSender;
 
 public class ChatHandler implements Listener
 {
