+++ src/main/java/com/festp/messaging/MessageSender.java
@@ -1,0 +1,8 @@
+package com.festp.messaging;
+
+import org.bukkit.entity.Player;
+
+public interface MessageSender
+{
+	public void sendRawJson(Player player, CharSequence rawJsonMessage);
+}
